+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Add featured packs tab
 - Add message to accounts tab with info on what account to use
+- Fix loading Forge loader on older MC versions